gw1h v0.1.0 - starts Guild Wars and GWToolbox
I'm running both Guild Wars and GWToolbox via wine. I made a simple bash
script in order to set the wine environment variables, start gw, and start
toolbox. That worked, but it bothered me that I had to select the account /
char in toolbox every time I launched. So, I started looking in to
extending my script to automate away the one button click. It quickly
became a mess that no longer worked. So, I decided to just write a go
program instead. So, gw1h is meant to automatically launch Guild Wars, find
the correct process id, and start toolbox with that pid so that I don't
have to click an extra button. v0.1.0 manages to do this, just barely. In
the future, gw1h should support running multiple Guild Wars instances, but
that's for another weekend.